Added Camera functionality to Level Template
The camera now follows the player around in the level template, which new levels will inherit. I had to look up the camera movement logic in HeartBeast's tutorial. Also, I had to move the Camera's child nodes, TopLeft and BottomRight, in order for the camera to be able to move.